Clean up and refactor mail sending code
APUB-398
